+++ test/reorderable_carousel_test.dart
@@ -117,4 +117,43 @@ void main() {
       expect(newIndex, 1);
     },
   );
+
+  testWidgets(
+    "Add item awaits",
+    (WidgetTester tester) async {
+      bool? addItemCalled;
+      bool? itemSelectedCalled;
+      await tester.pumpWidget(
+        MaterialApp(
+          home: Material(
+            child: ReorderableCarousel(
+              numItems: 1,
+              addItemAt: (int index) async {
+                await Future.delayed(Duration(seconds: 1));
+                addItemCalled = true;
+              },
+              itemBuilder: (double itemWidth, int index, bool isSelected) {
+                return Container(
+                  key: Key("Item $index"),
+                  height: 100,
+                  width: itemWidth,
+                );
+              },
+              onReorder: (int oldIdx, int newIdx) {},
+              onItemSelected: (_) {
+                expect(addItemCalled, true);
+                itemSelectedCalled = true;
+              },
+            ),
+          ),
+        ),
+      );
+      await tester.pumpAndSettle();
+
+      await tester.tap(find.byType(IconButton).first);
+      await tester.pumpAndSettle(const Duration(seconds: 20));
+
+      expect(itemSelectedCalled, true);
+    },
+  );
 }