+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.2.0 - 2021-03-31
+### Changed
+- addItemAt can now be used async. Meaning that onItemSelected won't be called until it completes
+
 ## 0.2.0-nullsafety - 2021-03-24
 ### Changed
 - Null safety